# Docs/openapi.stub.yaml
openapi: 3.0.3
info:
  title: Inabiya API (stub)
  version: 0.1.0-dev
  description: |
    Minimal OpenAPI stub for Soft Gift commerce + CMS (dev leftovers Wave 5).
    Not a full generated surface — expand when `@nestjs/swagger` is adopted.
    Payments remain mock (`PAYMENT_PROVIDER=mock`); Razorpay/S3/SMTP are post-dev.
servers:
  - url: http://127.0.0.1:4001/api/v1
    description: VPS loopback
paths:
...
  /catalog/products:
    get:
      summary: Published catalog list
      parameters:
        - in: query
          name: category
          schema: { type: string }
        - in: query
          name: hamper
          schema: { type: string, enum: ['0', '1'] }
      responses:
        '200':
          description: Product summaries
  /catalog/products/{slug}:
    get:
      summary: Published product detail (includes SEO + faqItems when set)
      parameters:
        - in: path
          name: slug
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Product DTO
        '404':
          description: Not found
  /admin/catalog/products/{id}:
    patch:
      summary: Update product (commerce admin) — SEO + FAQ fields
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Updated product
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT